+++ jsonpath/env.py
@@ -24,6 +24,7 @@ from .exceptions import JSONPathSyntaxError
 from .filter import UNDEFINED
 from .function_extensions import validate
 from .lex import Lexer
+from .match import JSONPathMatch
 from .match import NodeList
 from .parse import Parser
 from .path import CompoundJSONPath
@@ -36,7 +37,6 @@ from .token import Token
 
 if TYPE_CHECKING:
     from .match import FilterContextVars
-    from .match import JSONPathMatch
 
 
 class JSONPathEnvironment:
@@ -106,6 +106,7 @@ class JSONPathEnvironment:
     # Override these to customize path tokenization and parsing.
     lexer_class: Type[Lexer] = Lexer
     parser_class: Type[Parser] = Parser
+    match_class: Type[JSONPathMatch] = JSONPathMatch
 
     def __init__(self) -> None:  # noqa: D107
         self.lexer: Lexer = self.lexer_class(env=self)